Add missing tango icons
These icons are all referenced in the Mumble theme but were either
previously located in the Classic theme or were located there but not
aliased to at all (causing a bunch of icons not being found when running
Mumble - noticeable in the CertWizard).

Now that the Classic theme gets removed, it is time to add these icons
to this theme directly to have it all in a single place.

Note that these icons are licensed under CC-by-sa as stated here:
https://en.wikipedia.org/wiki/Tango_Desktop_Project